requirements.txt
customtkinter~=5.2.2
protobuf~=5.29.5
darkdetect==0.8.0
pillow==12.0.0

requests==2.32.5

google-generativeai==0.8.5
openai==0.27.10

pydantic==2.12.5
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.67.1
